looking good

+++ ios/GenQuizitive/GenQuizitive/QuestionService.swift
@@ -54,8 +54,8 @@ class GenQuizRound {
             myTotalIncorrect += q.myIncorrectCount
             friendTotalIncorrect += q.friendIncorrectCount
             
-            myTotalTime += q.myIncorrectCount * 20
-            friendTotalTime += q.friendIncorrectCount * 20
+            myTotalTime += Double(q.myIncorrectCount * 20)
+            friendTotalTime += Double(q.friendIncorrectCount * 20)
         }
     }
 }
